+++ src/ReceivingList/ExportSettingsModal/ExportSettingsModalContainer.js
@@ -1,11 +1,15 @@
 import { useCallback } from 'react';
 import PropTypes from 'prop-types';
-import noop from 'lodash/noop';
+
+import { useShowCallout } from '@folio/stripes-acq-components';
 
 import {
   EXPORT_FIELDS_PARAMS,
   EXPORT_SETTINGS_FIELDS,
 } from './constants';
+import {
+  usePiecesExportCSV,
+} from './hooks';
 import {
   getExportFields,
 } from './utils';
@@ -13,7 +17,14 @@ import ExportSettingsModal from './ExportSettingsModal';
 
 export const ExportSettingsModalContainer = ({
   onCancel,
+  query,
 }) => {
+  const showCallout = useShowCallout();
+  const {
+    runExportCSV,
+    isLoading,
+  } = usePiecesExportCSV();
+
   const initialValues = {
     [EXPORT_SETTINGS_FIELDS.exportTitleFields]: EXPORT_FIELDS_PARAMS.all,
     [EXPORT_SETTINGS_FIELDS.exportPieceFields]: EXPORT_FIELDS_PARAMS.all,
@@ -22,14 +33,19 @@ export const ExportSettingsModalContainer = ({
   const onSubmit = useCallback((formValues) => {
     const exportFields = getExportFields(formValues);
 
-    noop(exportFields);
-    onCancel();
-  }, [onCancel]);
+    showCallout({ messageId: 'ui-receiving.exportSettings.success' });
+    runExportCSV({
+      exportFields,
+      query,
+    })
+      .catch(() => showCallout({ messageId: 'ui-receiving.exportSettings.error', type: 'error' }))
+      .finally(onCancel);
+  }, [onCancel, runExportCSV, showCallout, query]);
 
   return (
     <ExportSettingsModal
       initialValues={initialValues}
-      isExporting={false}
+      isExporting={isLoading}
       onCancel={onCancel}
       onSubmit={onSubmit}
     />
@@ -38,4 +54,5 @@ export const ExportSettingsModalContainer = ({
 
 ExportSettingsModalContainer.propTypes = {
   onCancel: PropTypes.func.isRequired,
+  query: PropTypes.string.isRequired,
 };