+++ src/ReceivingList/ExportSettingsModal/utils/index.js
@@ -1,2 +1,5 @@
+export { createExportReport } from './createExportReport';
+export { getExportData } from './getExportData';
 export { getExportFields } from './getExportFields';
 export { isAllSelected } from './isAllSelected';
+export { mapUniqElements } from './mapUniqElements';