+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * ui-receiving: module warnings analysis. Refs UIREC-237.
 * Add "Export results (CSV)" action to receiving app. Refs UIREC-233.
 * Allow user to select data points for Export results to CSV. Refs UIREC-234.
+* Export pieces functionality - FE approach. Refs UIREC-235.
 
 ## [2.1.0](https://github.com/folio-org/ui-receiving/tree/v2.1.0) (2022-03-03)
 [Full Changelog](https://github.com/folio-org/ui-receiving/compare/v2.0.3...v2.1.0)